+++ reverse_engineering/SchemaCreator.js
@@ -19,7 +19,7 @@ const snippets = {
 	'completionObject': require(snippetsPath + 'completionObject.json'),
 };
 
-const helper = require('../helper/helper');
+const helper = require('../shared/propertiesHelper');
 
 module.exports = {
 	indices: [],
@@ -81,7 +81,7 @@ module.exports = {
 		};
 	},
 
-	getSchema(elasticMapping, sample) {
+	getSchema(elasticMapping, sample, fieldLevelConfig) {
 		let schema = this.getSchemaTemplate();
 		sample = sample || {};
 
@@ -90,6 +90,7 @@ module.exports = {
 			elasticMapping.properties,
 			sample._source,
 			elasticMapping.properties,
+			fieldLevelConfig,
 		);
 
 		if (elasticMapping.dynamic) {
@@ -99,7 +100,7 @@ module.exports = {
 		return schema;
 	},
 
-	getFields(properties, sample, mapping) {
+	getFields(properties, sample, mapping, fieldLevelConfig) {
 		let schema = {};
 
 		for (let fieldName in properties) {
@@ -109,13 +110,13 @@ module.exports = {
 				continue;
 			}
 
-			schema[fieldName] = this.getField(properties[fieldName], currentSample, mapping);
+			schema[fieldName] = this.getField(properties[fieldName], currentSample, mapping, fieldLevelConfig);
 		}
 
 		return schema;
 	},
 
-	getField(fieldData, sample, mapping) {
+	getField(fieldData, sample, mapping, fieldLevelConfig) {
 		let schema = {};
 
 		if (!fieldData) {
@@ -132,6 +133,7 @@ module.exports = {
 				fieldData.properties,
 				isArrayType && Array.isArray(sample) ? sample[0] : sample,
 				mapping,
+				fieldLevelConfig,
 			);
 
 			if (isArrayType) {
@@ -165,7 +167,7 @@ module.exports = {
 			schema = this.handleCompletionSnippet(schema);
 		}
 
-		schema = this.setProperties(schema, fieldData);
+		schema = this.setProperties(schema, fieldData, fieldLevelConfig);
 
 		return schema;
 	},
@@ -429,13 +431,14 @@ module.exports = {
 		return schema;
 	},
 
-	setProperties(schema, fieldData) {
+	setProperties(schema, fieldData, fieldLevelConfig) {
 		const properties = helper.getFieldProperties(
 			schema.type,
 			{ mode: fieldData.type, ...fieldData },
 			{
 				'stringfields': 'fields',
 			},
+			fieldLevelConfig,
 		);
 
 		for (let propName in properties) {
